Python scripts for receiving data
Added data_collector with the realtimeplotter.py for receiving data from the robot and being able to plot it in real time and/or save it. The data_extender.py script extends the data from the robot by results stemming from calculations from the model with the received data as input.